+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - `Innmind\Filesystem\Adapter::all()`
 - `Innmind\Filesystem\Chunk`
 - `Innmind\Filesystem\File\Content\Chunkable`
+- Possibility to use a `Innmind\Immutable\Set` of files inside a `Directory`
 
 ## 6.6.0 - 2023-09-16